adiciona verificação para ID na exclusão de tarefas, retornando erro 404 se ID não for fornecido

+++ Tarefas/views.py
@@ -70,7 +70,9 @@ class TarefaModelViewSet(ModelViewSet):
         
     def delete(self, request):
         id = request.GET.get('id')
-        taf = Tarefa.objects.get(id=id)
+        if not id:
+            return Response({'status':404, 'msg': 'bad request'})
+        taf = Tarefa.objects.filter(id=id)
         taf.delete()
         return Response({'status': 200, 'msg': 'Deleted!'})
